some logging progress, not really working :(

+++ flight_software/inc/delay.h
@@ -11,7 +11,8 @@ void TIM4_UPD_OVF_IRQHandler(void) __interrupt(23);
 void TIM4_Config(void);
 void delay_ms(uint16_t ms);
 void delay_us(uint16_t us);
-uint16_t millis(void);
+/** returns the number of milliseconds that have passed since boot  **/
+uint32_t millis(void);
 uint32_t micros(void);
 
 #endif //DELAY_H
